Fix: decimales rotos desde 1.13, ajuste manual y traslados de stock
Encontrado en la Fase 2 del plan de seguimiento (PLAN_SEGUIMIENTO.md), al
revisar los formularios de ajuste manual. Tres bugs reales, todos
reproducidos contra la base local antes de corregirlos:

1. utils/inventario.js (regresion de 3cbc1ad, en 1.13.0 y 1.14.0): el
   upsert usaba GREATEST($4, 0) sin cast. Postgres deducia $4 como INTEGER
   por el literal 0, y cualquier delta con decimales (Kilo / Metro, Fase
   11) fallaba con "sintaxis no valida para tipo integer". Afectaba a
   guias, devoluciones y transferencias. Ahora es $4::numeric.

2. POST /api/inventario (ajuste manual): era un INSERT directo. Desde el
   UNIQUE (almacen, producto, tipo) de 3cbc1ad, un segundo ajuste sobre un
   producto que ya tenia stock fallaba con "llave duplicada" (500). Ahora
   suma con ajustarInventario y valida almacen, tipo, cantidad > 0 con
   hasta 3 decimales, y decimales solo si la unidad lo permite.

3. POST /api/movimientos (traslado): pg devuelve NUMERIC como string, asi
   que la suma de filas CONCATENABA ("05.0003.000") y el chequeo de stock
   no se cumplia con 2 filas (NUEVO + DEVOLUCION). Un traslado de 100 con
   8 disponibles pasaba y creaba 92 en destino. Ahora el calculo va en
   milesimas enteras, el destino usa el mismo upsert y se aplican las
   mismas validaciones de cantidad y unidad.

Tambien:
- La validacion de cantidades pasa a utils/cantidad.js, compartida con
  guias.
- /api/productos e /api/inventario/disponible devuelven permite_decimal.
  Inventario.jsx y Movimientos.jsx usan step/min 0.001 solo para KG/M/M2/L.
- package.json: backend/*_log.txt fuera del instalador.

Probado por API (16 casos): 2do ajuste suma (7), negativos, 4 decimales,
decimal en producto entero, tipo invalido y almacen inexistente dan 400.
Traslado de 100 con 10 da 400, el de 9 toma de NUEVO y DEVOLUCION, 0.7 m
ok y 2.1 m con 2.05 da 400. Una guia de 2.5 m (antes 500) da 201.

+++ backend/src/utils/inventario.js
@@ -14,11 +14,14 @@ async function ajustarInventario(client, { almacenId, productoId, delta, descrip
   // para el mismo (almacen,producto,tipo) que TODAVIA no existia podian
   // pasar los dos por "no existe" y terminar creando 2 filas duplicadas en
   // vez de una sola. Con ON CONFLICT, Postgres serializa la operacion sola.
+  // $4::numeric explicito: sin el cast, GREATEST($4, 0) hace que Postgres
+  // deduzca $4 como INTEGER (por el literal 0) y cualquier delta con decimales
+  // (Kilo / Metro, Fase 11) falla con "sintaxis no valida para tipo integer".
   await conexion.query(
     `INSERT INTO inventario (almacen_id, producto_id, tipo, cantidad, descripcion)
-     VALUES ($1, $2, $3, GREATEST($4, 0), $5)
+     VALUES ($1, $2, $3, GREATEST($4::numeric, 0), $5)
      ON CONFLICT (almacen_id, producto_id, tipo)
-     DO UPDATE SET cantidad = GREATEST(inventario.cantidad + $4, 0)`,
+     DO UPDATE SET cantidad = GREATEST(inventario.cantidad + $4::numeric, 0)`,
     [almacenId, productoId, tipo, delta, descripcion || null]
   )
 }
